Added a splash screen >using native splash >changing the gradient colors in the login, sign up, forgot password to a solid color(deepPurple) >added dumbbell logo(assets/splash/dumbbell-white.png)
