+++ android/app/src/main/kotlin/com/arlonie/fitness_app/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.arlonie.fitness_app
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()
