+++ android/app/src/main/res/drawable-night-mdpi/android12splash.png
