+++ android/app/src/main/res/drawable-night-mdpi/splash.png
