+++ android/app/src/main/res/drawable-night-v21/launch_background.xml
