+++ android/app/src/main/res/drawable-night-xhdpi/splash.png
