+++ android/app/src/main/res/drawable-night-xxhdpi/splash.png
