+++ android/app/src/main/res/drawable-xhdpi/android12splash.png
