+++ lib/main.dart
@@ -1,9 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
-import 'package:firebase_auth/firebase_auth.dart';
 import 'firebase_options.dart';
-import 'screens/login_screen.dart';
-import 'screens/home_screen.dart';
+import 'screens/wrapper.dart'; // 👈 add this
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -20,30 +18,7 @@ class MyApp extends StatelessWidget {
       debugShowCheckedModeBanner: false,
       title: 'Fitness App',
       theme: ThemeData(primarySwatch: Colors.blue),
-      home: const Wrapper(), // 👈 check if logged in or not
-    );
-  }
-}
-
-class Wrapper extends StatelessWidget {
-  const Wrapper({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return StreamBuilder<User?>(
-      stream: FirebaseAuth.instance
-          .authStateChanges(), // 👈 listens to login state
-      builder: (context, snapshot) {
-        if (snapshot.connectionState == ConnectionState.waiting) {
-          return const Scaffold(
-            body: Center(child: CircularProgressIndicator()),
-          );
-        } else if (snapshot.hasData) {
-          return const HomeScreen(); // 👈 user already logged in
-        } else {
-          return const LoginScreen(); // 👈 not logged in
-        }
-      },
+      home: const Wrapper(), // 👈 this handles login state
     );
   }
 }
