+++ lib/screens/login_screen.dart
@@ -106,11 +106,13 @@ class _LoginScreenState extends State<LoginScreen> {
       body: Container(
         width: double.infinity,
         decoration: const BoxDecoration(
-          gradient: LinearGradient(
-            colors: [Colors.deepPurple, Colors.pinkAccent],
-            begin: Alignment.topLeft,
-            end: Alignment.bottomRight,
-          ),
+          color: Colors.deepPurple,
+          // gradient: LinearGradient(
+          //   // colors: [Colors.deepPurple, Colors.pinkAccent],
+          //   colors: [Colors.deepPurple],
+          //   begin: Alignment.topLeft,
+          //   end: Alignment.bottomRight,
+          // ),
         ),
         child: SafeArea(
           child: Center(
@@ -129,13 +131,13 @@ class _LoginScreenState extends State<LoginScreen> {
                     builder: (context, value, child) {
                       return Transform.scale(scale: value, child: child);
                     },
-                    child: const Icon(
-                      Icons.fitness_center,
-                      size: 100,
-                      color: Colors.white,
+                    child: Image.asset(
+                      'assets/splash/dumbbell-white.png',
+                      width: 100, // adjust size here
+                      height: 100, // adjust size here
+                      fit: BoxFit.contain,
                     ),
                   ),
-                  const SizedBox(height: 20),
 
                   // Heading Text
                   const Text(
