+++ lib/screens/signup_screen.dart
@@ -141,11 +141,12 @@ class _SignUpScreenState extends State<SignUpScreen> {
       body: Container(
         width: double.infinity,
         decoration: const BoxDecoration(
-          gradient: LinearGradient(
-            colors: [Colors.deepPurple, Colors.pinkAccent],
-            begin: Alignment.topLeft,
-            end: Alignment.bottomRight,
-          ),
+          color: Colors.deepPurple,
+          // gradient: LinearGradient(
+          //   colors: [Colors.deepPurple, Colors.pinkAccent],
+          //   begin: Alignment.topLeft,
+          //   end: Alignment.bottomRight,
+          // ),
         ),
         child: SafeArea(
           child: Center(
@@ -157,12 +158,14 @@ class _SignUpScreenState extends State<SignUpScreen> {
                   TweenAnimationBuilder<double>(
                     tween: Tween(begin: 0, end: 1),
                     duration: const Duration(milliseconds: 800),
-                    builder: (context, value, child) =>
-                        Transform.scale(scale: value, child: child),
-                    child: const Icon(
-                      Icons.fitness_center,
-                      size: 100,
-                      color: Colors.white,
+                    builder: (context, value, child) {
+                      return Transform.scale(scale: value, child: child);
+                    },
+                    child: Image.asset(
+                      'assets/splash/dumbbell-white.png',
+                      width: 100, // adjust size here
+                      height: 100, // adjust size here
+                      fit: BoxFit.contain,
                     ),
                   ),
                   const SizedBox(height: 16),
