+++ lib/screens/wrapper.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'home_screen.dart';
+import 'login_screen.dart';
+
+class Wrapper extends StatelessWidget {
+  const Wrapper({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return StreamBuilder<User?>(
+      stream: FirebaseAuth.instance.authStateChanges(),
+      builder: (context, snapshot) {
+        // While checking auth state → show loading screen
+        if (snapshot.connectionState == ConnectionState.waiting) {
+          return const Scaffold(
+            body: Center(child: CircularProgressIndicator()),
+          );
+        }
+
+        // If logged in → go to home
+        if (snapshot.hasData) {
+          return const HomeScreen();
+        }
+
+        // If not logged in → go to login
+        return const LoginScreen();
+      },
+    );
+  }
+}
